small fixes for fonts on mac and win32; changed Thread::setPriority to return a bool; added a File::getVolumeTotalSize() method; added a freetype path to the linux makefile; tweaked the jucer to make cached image variable names more readable; tweaked the hit-spots in the resizable borders of very small windows; moved ResizableWindow background painting into the LookAndFeel; made the broadcast flag of a DatagramSocket an optional constructor parameter;

+++ src/juce_appframework/audio/audio_file_formats/juce_AudioFormat.cpp
@@ -85,7 +85,7 @@ bool AudioFormatReader::read (int** destSamples,
                        startSampleInSource, numSamplesToRead))
         return false;
 
-    if (numDestChannels > numChannels)
+    if (numDestChannels > (int) numChannels)
     {
         if (fillLeftoverChannelsWithCopies)
         {
